+++ realm/realm-library/src/syncIntegrationTest/java/io/realm/SyncedRealmIntegrationTests.java
@@ -60,7 +60,6 @@ public class SyncedRealmIntegrationTests extends StandardIntegrationTest {
 
         SyncConfiguration config = user.createConfiguration(Constants.USER_REALM)
                 .schema(StringOnly.class)
-                .fullSynchronization()
                 .sessionStopPolicy(OsRealmConfig.SyncSessionStopPolicy.IMMEDIATELY)
                 .build();
 
@@ -85,7 +84,6 @@ public class SyncedRealmIntegrationTests extends StandardIntegrationTest {
 
         user = SyncUser.logIn(SyncCredentials.usernamePassword(username, password, false), Constants.AUTH_URL);
         SyncConfiguration config2 = user.createConfiguration(Constants.USER_REALM)
-                .fullSynchronization()
                 .schema(StringOnly.class)
                 .build();
 
@@ -102,7 +100,6 @@ public class SyncedRealmIntegrationTests extends StandardIntegrationTest {
     public void waitForInitialRemoteData_mainThreadThrows() {
         final SyncUser user = SyncTestUtils.createTestUser(Constants.AUTH_URL);
         SyncConfiguration config = user.createConfiguration(Constants.USER_REALM)
-                .fullSynchronization()
                 .waitForInitialRemoteData()
                 .build();
 
@@ -126,7 +123,6 @@ public class SyncedRealmIntegrationTests extends StandardIntegrationTest {
 
         // 1. Copy a valid Realm to the server (and pray it does it within 10 seconds)
         final SyncConfiguration configOld = configurationFactory.createSyncConfigurationBuilder(user, Constants.USER_REALM)
-                .fullSynchronization()
                 .schema(StringOnly.class)
                 .sessionStopPolicy(OsRealmConfig.SyncSessionStopPolicy.IMMEDIATELY)
                 .build();
@@ -148,7 +144,6 @@ public class SyncedRealmIntegrationTests extends StandardIntegrationTest {
         user = SyncUser.logIn(SyncCredentials.usernamePassword(username, password), Constants.AUTH_URL);
         SyncConfiguration config = user.createConfiguration(Constants.USER_REALM)
                 .name("newRealm")
-                .fullSynchronization()
                 .schema(StringOnly.class)
                 .waitForInitialRemoteData()
                 .build();
@@ -248,7 +243,6 @@ public class SyncedRealmIntegrationTests extends StandardIntegrationTest {
 
         // 1. Copy a valid Realm to the server (and pray it does it within 10 seconds)
         final SyncConfiguration configOld = configurationFactory.createSyncConfigurationBuilder(user, Constants.USER_REALM)
-                .fullSynchronization()
                 .schema(StringOnly.class)
                 .build();
         Realm realm = Realm.getInstance(configOld);
@@ -269,7 +263,6 @@ public class SyncedRealmIntegrationTests extends StandardIntegrationTest {
         user = SyncUser.logIn(SyncCredentials.usernamePassword(username, password, false), Constants.AUTH_URL);
         final SyncConfiguration configNew = configurationFactory.createSyncConfigurationBuilder(user, Constants.USER_REALM)
                 .name("newRealm")
-                .fullSynchronization()
                 .waitForInitialRemoteData()
                 .readOnly()
                 .schema(StringOnly.class)
@@ -450,34 +443,6 @@ public class SyncedRealmIntegrationTests extends StandardIntegrationTest {
         looperThread.closeAfterTest(realm);
     }
 
-
-    /**
-     * Tests https://github.com/realm/realm-java/issues/6235
-     * This checks that the INITIAL callback is called for query-based notifications even when
-     * the device is offline.
-     */
-    @Test
-    @RunTestInLooperThread
-    public void listenersTriggerWhenOffline() {
-        SyncUser user = SyncTestUtils.createTestUser(); // Creating a fake user will make it behave as "offline"
-        String url = "http://foo.com/offlineListeners";
-        SyncConfiguration config = configurationFactory.createSyncConfigurationBuilder(user, url)
-                .build();
-        Realm realm = Realm.getInstance(config);
-        looperThread.closeAfterTest(realm);
-
-        RealmResults<AllTypes> results = realm.where(AllTypes.class).findAllAsync();
-
-        looperThread.keepStrongReference(results);
-        results.addChangeListener((objects, changeSet) -> {
-            if(changeSet.getState() == OrderedCollectionChangeSet.State.INITIAL) {
-                assertTrue(results.isLoaded());
-                assertFalse(changeSet.isCompleteResult());
-                looperThread.testComplete();
-            }
-        });
-    }
-
     @Test
     @RunTestInLooperThread
     public void progressListenersWorkWhenUsingWaitForInitialRemoteData() throws InterruptedException {
@@ -487,7 +452,6 @@ public class SyncedRealmIntegrationTests extends StandardIntegrationTest {
 
         // 1. Copy a valid Realm to the server (and pray it does it within 10 seconds)
         final SyncConfiguration configOld = configurationFactory.createSyncConfigurationBuilder(user, Constants.USER_REALM)
-                .fullSynchronization()
                 .schema(StringOnly.class)
                 .sessionStopPolicy(OsRealmConfig.SyncSessionStopPolicy.IMMEDIATELY)
                 .build();
@@ -510,7 +474,6 @@ public class SyncedRealmIntegrationTests extends StandardIntegrationTest {
         user = SyncUser.logIn(SyncCredentials.usernamePassword(username, password), Constants.AUTH_URL);
         SyncConfiguration config = user.createConfiguration(Constants.USER_REALM)
                 .name("newRealm")
-                .fullSynchronization()
                 .schema(StringOnly.class)
                 .waitForInitialRemoteData()
                 .build();