+++ realm/realm-library/src/syncTestUtils/java/io/realm/objectserver/utils/UserFactory.java
@@ -51,7 +51,6 @@ public class UserFactory {
     private static synchronized void initFactory(boolean forceReset) {
         if (configuration == null || forceReset) {
             RealmConfiguration.Builder builder = new RealmConfiguration.Builder().name("user-factory.realm");
-            ObjectServerFacade.getSyncFacadeIfPossible().addSupportForObjectLevelPermissions(builder);
             configuration = builder.build();
         }
     }